+++ txpipe/source_selection/__init__.py
@@ -3,4 +3,5 @@ from .hsc import TXSourceSelectorHSC
 from .lensfit import TXSourceSelectorLensfit
 from .metacal import TXSourceSelectorMetacal
 from .metadetect import TXSourceSelectorMetadetect
+from .dp2 import TXSourceSelectorMetadetectDP2
 from .tomography import TXSourceTomography
